SAML authentication
I added support for SAML authentication where the user logs in Smart Office in the form domain\sAMAccountName, but in MNS150 it's still the form USID. In IPA, the Identity is domain\sAMAccountName, the Actor is USID, and the Actor-Identity connects them together. Also, I am getting the firstname and lastname from LDAP now to avoid incorrect name splits, and I am getting the email address from LDAP for the Create Actor; then CRS111 takes over.